+++ packages/drivers/odsp-driver/src/epochTracker.ts
@@ -273,7 +273,7 @@ export class EpochTracker implements IPersistedFileCache {
             // If it was categorized as epoch error but the epoch returned in response matches with the client epoch
             // then it was coherency 409, so rethrow it as throttling error so that it can retried. Default throttling
             // time is 1s.
-            // this.logger.sendErrorEvent({ eventName: "Coherency409" }, error);
+            this.logger.sendErrorEvent({ eventName: "Coherency409" }, error);
             throw new ThrottlingError(error.errorMessage ?? "Coherency409", 1, { [Odsp409Error]: true });
         }
     }
