+++ packages/runtime/container-runtime/src/index.ts
@@ -7,7 +7,6 @@ export * from "./containerRuntime";
 export * from "./deltaScheduler";
 export * from "./dataStoreRegistry";
 export * from "./pendingStateManager";
-export * from "./runWhileConnectedCoordinator";
 export * from "./summarizer";
 export * from "./summarizerTypes";
 export * from "./summaryCollection";
